# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Channel Copier API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  phone?: string | null;
}

export interface SendCodeInput {
  /** Phone number with country code e.g. +40712345678 */
  phone: string;
}

export interface SendCodeResult {
  phoneCodeHash: string;
}

export interface SignInInput {
  phone: string;
  phoneCodeHash: string;
  code: string;
}

export interface SignInResult {
  authenticated: boolean;
  requires2fa: boolean;
  /** @nullable */
  phone?: string | null;
}

export interface SignIn2faInput {
  password: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  stopped: 'stopped',
  failed: 'failed',
} as const;

export interface Job {
  id: string;
  sourceChannel: string;
  destChannel: string;
  status: JobStatus;
  totalPosts: number;
  copiedPosts: number;
  failedPosts: number;
  createdAt: string;
  /** @nullable */
  finishedAt?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface JobInput {
  /** Source channel username or invite link */
  sourceChannel: string;
  /** Destination channel username or invite link */
  destChannel: string;
}