# lib/api-zod/src/generated/types/authStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Channel Copier API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  phone?: string | null;
}