# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Channel Copier API
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from './jobStatus';

export interface Job {
  id: string;
  sourceChannel: string;
  destChannel: string;
  status: JobStatus;
  totalPosts: number;
  copiedPosts: number;
  failedPosts: number;
  createdAt: string;
  /** @nullable */
  finishedAt?: string | null;
  /** @nullable */
  error?: string | null;
}