# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Channel Copier API
 * OpenAPI spec version: 0.1.0
 */

export interface JobInput {
  /** Source channel username or invite link */
  sourceChannel: string;
  /** Destination channel username or invite link */
  destChannel: string;
}