# lib/api-zod/src/generated/types/sendCodeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Channel Copier API
 * OpenAPI spec version: 0.1.0
 */

export interface SendCodeInput {
  /** Phone number with country code e.g. +40712345678 */
  phone: string;
}