# lib/api-zod/src/generated/types/signInInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Channel Copier API
 * OpenAPI spec version: 0.1.0
 */

export interface SignInInput {
  phone: string;
  phoneCodeHash: string;
  code: string;
}