# lib/api-zod/src/generated/types/signInResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Channel Copier API
 * OpenAPI spec version: 0.1.0
 */

export interface SignInResult {
  authenticated: boolean;
  requires2fa: boolean;
  /** @nullable */
  phone?: string | null;
}